Move MaybeTlsStream from sinks/util/tls into tls
